pyproject.toml
[project]
name = "facttrack"
version = "0.1.0"
description = "East Texas landwork automation pipeline — public-records ingestion + curative triage."
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "playwright>=1.45",
    "psycopg2-binary>=2.9",
    "pydantic>=2.7",
    "python-dateutil>=2.9",
    "scipy>=1.13",
    "openpyxl>=3.1",
    "jinja2>=3.1",
    "weasyprint>=62",
    "folium>=0.16",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "tenacity>=8.2",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "black>=24.4",
    "mypy>=1.10",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]
